requirements.txt
Flask>=2.2.2
Flask-Login>=0.6.2
Werkzeug>=2.1.2
Pillow>=9.4.0
opencv-python>=4.7.0
numpy>=1.24.0
torch>=2.0.0
